requirements.txt
fastapi
pydantic
pydantic[email]
psycopg2
sqlalchemy
python-dotenv
uvicorn
passlib
bcrypt
python-jose
pydantic-settings
alembic
httpx
pytest